move installation stuff into app folder
