+++ backend/src/main/java/spectacular/backend/WebConfiguration.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import spectacular.backend.installation.InstallationIdInterceptor;
+import spectacular.backend.app.InstallationIdInterceptor;
 
 @Configuration
 public class WebConfiguration implements WebMvcConfigurer {
