--- backend/src/main/java/spectacular/backend/installation/InstallationError.java
@@ -1,13 +1,0 @@
-package spectacular.backend.installation;
-
-public class InstallationError {
-  private final String error;
-
-  public InstallationError(String error) {
-    this.error = error;
-  }
-
-  public String getError() {
-    return error;
-  }
-}
